--- src/components/formAction/Form.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Form() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Form
