+++ src/containers/Main.js
@@ -16,6 +16,7 @@ import ScrollToTopButton from "./topbutton/Top";
 import Twitter from "./twitter-embed/twitter";
 import Profile from "./profile/Profile";
 import SplashScreen from "./splashScreen/SplashScreen";
+import Form from "./form/Form";
 import {splashScreen} from "../portfolio";
 import {StyleProvider} from "../contexts/StyleContext";
 import {useLocalStorage} from "../hooks/useLocalStorage";
@@ -63,6 +64,7 @@ const Main = () => {
             <Talks />
             <Twitter />
             <Podcast />
+            <Form />
             <Profile />
             <Footer />
             <ScrollToTopButton />
